Rename toplevel to responses
Moved all non-response models/classes to optimade_json.
This mainly includes Warnings and ResponseMeta and all their
used/internal models.
